recipe_scraper/requirements.txt
flask>=3.0
recipe-scrapers>=15.0
pywebview>=5.0
requests>=2.31
